requirements.txt
asgiref==3.4.1
Django==3.2.8
django-ckeditor==6.1.0
django-jet-reboot==1.3.3
django-js-asset==1.2.2
django-recaptcha==3.0.0
Pillow==8.3.2
python-decouple==3.8
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
whitenoise==6.6.0
